all: Rework of ReadPlanExecutor
This commit add new definition of ReadPlan
and changes the implementation of ReadPlanExecutor.

The read plan uses new concept of waves.
Each wave starts read operation on some of the chunkservers.
If any of the operations will fail or wave will timeout,
then next wave starts. Whole read operation is completed when
we have all the required data or they can be reconstructed.
The wave approach allows for much greater flexibility
when dealing with erasure code chunks.
ReadPlanExecutor is rewritten to match ReadPlan,
with tiny reduction in number of lambda functions used.

Change-Id: I7a27b0da7b61b37fbb47cd6a87131e843a48fb02