master: Enable parsing erasure code goals
This commit enables parsing erasure code goal types.
Valid format is as follows:
 1 erasure_goal: $ec(3,5) {A B C}

It consists of erasure code token (ec) and 2 integer values,
which stand for quantities of data parts and parity parts respectively.
Additionally, label for each part can be specified as in any other
goal type.

Valid range of data parts is [2, 32].
Valid range of parity parts is [1, 32].

Change-Id: I886740aa5043cc208387c1a610a49765df70ee70