tools: Add displaying part numbers for ec chunks
This commit adds part number information for chunks stored
in erasure code goal. New display format is:

a/file:
	chunk 0: 0000000000000001_00000001 / (id:1 ver:1)
		copy 1: 192.168.88.202:10021:l0 part 4/5 of ec(3,2)
		copy 2: 192.168.88.202:10022:l1 part 2/5 of ec(3,2)
		copy 3: 192.168.88.202:10023:l2 part 5/5 of ec(3,2)
		copy 4: 192.168.88.202:10023:l4 part 2/3 of xor2
		copy 5: 192.168.88.202:10023:l2 part 1/3 of xor2
		copy 6: 192.168.88.202:10023:l5 part 3/3 of xor2

Please note that format of xor goals is changed as well, so grep formulas
in existing scripts would need to be amended. It was done in order to unify
printing format of xors, erasure codes and any future slice types.

Change-Id: Id21e414d1ac4c31c8eadb12ef7438b91092c0664