cmake: Add CXXDebug build type
This commit adds CXXDebug build type to cmake. This build type passes
-D_GLIBCXX_DEBUG directive to g++ compiler, which enables debug version
of standard C++ library.

If boost library is to be used with _GLIBCXX_DEBUG flag, it should be
compiled with -D_GLIBCXX_DEBUG as well to ensure matching ABI.
Therefore, using CMAKE_BUILD_TYPE=CxxDebug is advised to be followed
by ENABLE_POLONAISE=0.

Change-Id: Ice86bca84bf9291b75c1e9557486092316005e6c