master: Prioritize data parts before parity on chunk creation
This commit makes master choose data parts before parity when there are
not enough chunkservers to fulfill a goal and partial write must occur.
Thanks to that, data parts will always be available after creation
and no recovery would be needed to perform read operations.

This commit comes with a test which ensures that data parts are always
chosen before parity ones when there are not enough chunkservers
to fulfill goal requirements.

Change-Id: I8ebcddf6bf3f967f6868d9131de71bb2365115cc