master: Change trash emptying conditions
This commit changes conditions of removing files from trash.
From now on, access time (instead of creation time) of a file
is updated when it is trashed. Periodic trash collector checks
if file access time and its trashtime value meet the conditions
of being purged.

Thanks to this change, files with atime/mtime/ctime values
set to future would not reside in trash for a long period of time.

Closes #288

Change-Id: I35c071ffac1a56875c8a016924fecd0bc5e3f6b7