master: Remove keeping zombie servers in queues
This commit removes keeping zombie servers in separate queues.
New implementation of chunkserver database ensures that
information on disconnected chunkserver is kept in there
as well, so it is no longer needed to maintain separate
data structures for recently disconnected servers.
Instead, server's metadata can be safely removed the moment
it disconnects.

Change-Id: Ibd861416411f4a98ccf1171eb8e1dba346a5f2b0