Resolve 'mfsmaster' to connect CGI server to master
Currently, the CGI server uses the hard-coded loopback IP for the master host
without an option to change this address.

Other parts of LizardFS (e.g. shadow master) try to resolve 'mfsmaster' when
connecting to the master server. I reproduced the same behavior for the CGI
server. If 'mfsmaster' cannot be resolved, it will fall back to the loopback
IP, which ensures backwards-compatibility.