doc: Update documentation to include DISTANCE_PENALTY
This patch-set adds a new configuration option that allows the admin
to set a distance penalty for each chunkserver.  This penalty will be
added to *any* distance calculations made for the chunkserver, which
(depending on its value and whether my previous distance calculation
patch has been applied) will make the chunkserver a less likely candidate
for reads.

This patch updates the documentation to show the DISTANCE_PENALTY
configuration option.

Signed-off-by: Jonathan Dieter <jdieter@lesbg.com>