master: Add rebalancing of chunk parts from servers with same ip
This commit adds handling of chunk parts residing
on chunkservers with the same ip.

First of all there is extra step in chunkloop for moving chunk parts
to decrease number of parts with the same ip. Secondly rebalance based
on free space is limited, so it can't increase number of parts with the same ip.

Change-Id: Ia0247a7d39f991c235f282a1b0ba278e5eba267a