tests: Remove ramdisk usage from test_build_lizardfs
This commit drops USE_RAMDISK=YES option from a test that builds lizardfs
from source, because it's space-consuming. LizardFS source with external
dependencies takes up a lot of space, as does compiling it all.
2GB ramdisk proved to be not enough to handle this test, as it keeps
failing with ENOSPC recently. There is no valid reason for this test
to require a ramdisk, so it is removed. Now this test can be run on
test machines even if their ramdisk is <= 2GB in size.

Change-Id: I4b820937237bfd51f02a846d047d4484791a7413