Install proper manpage for mfsmount3 instead of symlink
Since it is possible to install only the client3 version, we do not want
the manpage to only be a symlink to the version using fuse2, but instead
we replace strings and generate a separate stand-alone manpage.

This also fixes the issue with symlinks in post-install scripts failing
when installed manpages are compressed.

Change-Id: Ief5fdd0785ed7d3d7e5a098587dad25cb7b4316a