mount: Fix listing of files from DirEntryCache in client's readdir
Fixes DirEntryCache in client's readdir returning wrong dirEntries
(from different dir or context).

The problem was present only when multiple threads (best seen with 4+)
were concurrently reading the same large collection of directories
with lots of files in it (for sure >100 in a single directory
was necessary for the directory to list some invalid file(s)).

Fixes #780

Change-Id: I823ce20aed06ac0d0a739f8954584dbc587a7fd8