tests: Add valgrind suppressions for external errors
Fixes suppression for uninitialized pwrite64 syscall parameter
in BerkeleyDB to cover all instances of this error.
Adds suppression for uninitialized values used in old LizardFS
version (3.11) which is used in one test.

This commit increases minimum required valgrind version for tests
to 3.15.0. This is because of using source-code-lines-based
suppressions for LizardFS 3.11 errors, which are available since
this version.

Change-Id: Ieff1bcb84cde5ceaf4903da62b9052ec00237a4e