tests: Fix ShortSystemTests.test_acl_cache
Fuse2 and 3 seem to work a little bit different under the hood.
In fuse3, during the test, while accessing single files in a directory,
their parent directory is being accessed by fuse and thus added to the cache,
and in fuse2 - not.
Extra config added in this commit caches the files' parent directory so that
it's being accessed by fuse only once, both in fuse2 and fuse3. The test
can still focus on testing whether acl_cache works properly, and it
does.

Change-Id: I525c3fc3c169a996222f7199637b37a61f790e4a