tests: Enable test timeouts to be machine-dependent
Adds the possibility to specify test timeout multiplier via
LIZARDFS_TEST_TIMEOUT_MULTIPLIER environment variable.
Its value can be floating-point. The total computed timeout
is then rounded down to a whole seconds value.

The purpose of introduction of this mutliplier is to make it possible
to relate a test's timeout to the performance of a machine running it.

The LIZARDFS_TEST_TIMEOUT_MULTIPLIER value must be set prior to
starting a test and changes to its value done while a test is running
will not change the timeout value.

Change-Id: I82577bbfc707add632a6ea95523b2c24ad9d65ca