cmake: Force building of static libs
The previous code base was incompatible to the cmake variable
setting '-DBUILD_SHARED_LIBS=YES', which didn't lead to build
failures, but produced an broken install, because installed
binaries were linked against shared libs which weren't installed.

This might not be an ideal solution, but it fixes the problem for now.
This commit fixes #880 by forcing all internal libraries to be
  STATIC.