mount: Fix updating user groups in master after its restart
The bug originated from clearing mount's group cache on master disconnected
event and using the same cache to register user's groups in the master later on.
The cache was already empty when attempting the registration so there were no
groups to tell master about so the operation performed on the mounted filesystem
failed.

This fix changes the registration part in the client so that it uses client
context to populate the cache and update credentials (groups) in master if
the group cache is empty, for example due to disconnection from master.

Context references were made non-const because updating groups in group cache
requires assigning a new gid value in the client context.

Fixes bug introduced by 7e7ad72

Change-Id: Idac2ebc74e73ca306462e07bd3b943c2efa73a0d