tests: Improve mounting functions in framework
Enables tests to change mount command when remounting.
Previously mount command was set only before setting up of
the in-test LizardFS infrastructure and couldn't be changed when
remounting.
Now the new mount command to use can be set in lizardfs_mount_start
function.

Unmounting function now waits for mount to be actually closed before
returning.
The previous behaviour (returning without waiting) is still preserved
in "async" version of this function.

Change-Id: I32b717bfe1c8a22b721b201f7e8756843bc57742