utils: Add sequence diagram generator for the protocol
When there is the need to explain the LizardFS protocol it gets
complex and extensive. It would be appropriate to have a graphical
representation for a better understanding. Sequence diagrams are
a common tool for describing interactions between elements of
the system architecture.

The new script parses the `MFSCommunication.h` file and produces
a declarative description that can be fed to a sequence diagram
generator, e.g., https://sequencediagram.org/

The script can receive arguments for two types of filters:
- participants : e.g., PARTICIPANTS='AN|CL|CS|MA|ML|TS'
- actions      : e.g., ACTIONS='CLTOMA|GET_CHUNK|WRITE_END'

Co-authored-by: guillex <guillex@lizardfs.org>

Change-Id: Id358d1fa7a3d46bb91525ae0f0016906340ebf81