tests: Force order of replication correctness checks based on file sizes
Upgrade tests generated files of various sizes:
2G 1G 100M 65M 64M 63M 55M 10M

Replication correctness timeouts are based on size of file replicated.
Before this commit the order of replication correctness checks were
random, after this commit the order is forced to be descending in file
size.

This approach ensures that short timeouts are used after long timeouts,
so replication of large file (which can be slow) won't impede on
replication correctness check of small files.

Change-Id: I8ecbd26695dd9a00b23449dced8a59e547da224c