tests: Add optional parallel-tests-dispatcher service
The parallel execution of the tests has been using a partitioning
algorithm to distribute the tests between the workers. This algorithm
depends on the expected time for the tests and the quantity
of available workers. It requires keeping those values updated and
could be error-prone if the deployment conditions change in the testing
infrastructure.

This commit adds a new approach using a web service for dispatching the
tests to the different workers. It doesn't rely on the number of workers
nor on the expected time. However, best performance should be
achieved if the provided list of tests is sorted in descending order
on the expected time.

The dispatcher uses Flask for the web service. It also has a Dockerfile
and a docker-compose.yml configuration file for easing the deployment.

Co-authored-by: guillex <guillex@lizardfs.org>
Change-Id: Icbb611992c2c9aa91c3d0480e514894aa5d4b934