mount: Improve readahead advisor feeding
The feed function was receiving the offset and size previously aligned
to match the memory layout of the chunks. This causes that, with
readings using very small block sizes (<= 16KB), the offset will be
repeated between two calls,  leading to an incorrect judgment as
random reading and a readahead window shrinking, even if the FUSE
reading was sequential.

This commit uses the original values from FUSE to feed the advisor,
instead of the aligned ones. The commit only affects the readahead
window size, not the data requested to the chunkserver.

Co-authored-by: Baldor <baldor@lizardfs.org>
Co-authored-by: Antuan <antuan@lizardfs.org>

Change-Id: I6d3d9ad2effd24e35a7fcca00cf2c1bab0d4e0c1