tests: Fix sed pattern for delete
A proper way to change the delimiter on the 'd' command to sed is to
escape it the first time is used. As the content of '$line' is expected
to include slash, the delimiter for sed was changed, but was missing the
escape character needed for 'd' (yet not needed for 's' :/).

Change-Id: I46cc4fe5cc7fa372be2b4f91836b66663fc6ab05