mount: Fix GID printing in the oplog
The LizardClient::Context can use the gid field as an index in the
GroupCache instead of the actual group number. The oplog was printing
this value without decoding it.

This commit fixes the oplog_printf function to decode the gid if
needed and adds a new test to be sure that oplog is printing the
correct id.

Change-Id: Idb45976d2d7d03f367739ee12548e3fe1ec93e1f